完成企业会员管理功能

+++ src/main/java/com/yida/boottracer/domain/SysMember.java
@@ -45,7 +45,7 @@ import net.minidev.json.annotate.JsonIgnore;
 public class SysMember extends AuditModel implements java.io.Serializable
 {
 
-	private int id;
+	private long id;
 	private DictCommon dictCompanyType;
 	private DictCommon dictIndustry;
 	private DictRegion dictRegion;
@@ -82,12 +82,12 @@ public class SysMember extends AuditModel implements java.io.Serializable
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "ID", unique = true, nullable = false)
-	public int getId()
+	public long getId()
 	{
 		return this.id;
 	}
 
-	public void setId(int id)
+	public void setId(long id)
 	{
 		this.id = id;
 	}
@@ -411,7 +411,7 @@ public class SysMember extends AuditModel implements java.io.Serializable
 	@Transient
 	public String getDisplayId() 
 	{
-		return StringUtils.leftPad(Integer.toString(this.getId()), 5, "0");
+		return StringUtils.leftPad(Long.toString(this.getId()), 5, "0");
 	}
 	
 	protected void setDisplayId(String v)
