+++ src/main/java/com/yida/boottracer/repo/SysMemberRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.yida.boottracer.domain.SysMember;
 
 @Repository
-public interface SysMemberRepository extends JpaRepository<SysMember, Integer>
+public interface SysMemberRepository extends JpaRepository<SysMember, Long>
 {
 	@Query("SELECT COUNT(d) FROM SysMember d WHERE d.dictMemberType.id=?1")
 	int myCountBydictCompanyType(int typeId);
